requirements.txt
# requirements.txt
# python -m playwright install chromium
# python -m playwright install --with-deps chromium

aiogram==3.13.1
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
playwright==1.47.0
tqdm==4.66.5
# python -m playwright install chromium

